documentation-template
## What's the purpose of this pull request?
We believe that documentation is essential for an educated decision flow — for internal and external developers. This is an invite to share the whys and make everybody an owner of the motivation and possibilities.

Today our technical writing team needs to investigate the technologies we use to create features, components, protocols, and etc. It takes time and VTEX's TW team is small by now.

We are also nearby the Fast Store open beta reality and we need to provide independence for our partner's developers to work with the new workflow.

## How it works?
It's a template with tips about what to write in each section. The main goal is to make us all go beyond the descriptions and start to explain more.

## How to test it?
Use it and send feedback to @VictoriaMolgado (@molgado on Slack).
- Does it fit in your daily work?
- What would you add or remove?

## References
Nothing by now, let's build this section together.
Join the discussion:
https://vtex.slack.com/archives/C0164SHMAV9/p1602253676083400